mother/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1
redis==5.0.6
aioredis==2.0.1
httpx==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pydantic-settings==2.3.0
websockets==12.0
